Workaround to add android.jar for transform
Close #2703

While javassit manipulating the method, it requires all of the classed
used in the method can be found in the class pool. But android.jar will
not be passed to the transform -- It is not in any of
QualifiedContent.Scope definitions.

We use reflection to get the AndroidBuilder instance from the android
plugin passed from gradle. AndroidBuilder.getBootClasspathAsStrings has
all the jar we need.